6/22 hp bar

+++ endless project/Assets/Script/Object/Player/Player.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 
 public class Player : ObjectManager
 {
-    public int dashConstant;
-    public int dashSpeed;
+    public float dashConstant;
+    public float dashSpeed;
 }
